+++ riseup/package.json
@@ -13,6 +13,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
     "react-router-dom": "5.0.1",
     "react-scripts": "^5.0.0",
     "styled-components": "^5.3.3",
