+++ riseup/src/App.js
@@ -1,7 +1,9 @@
 // import sunbeams from './sunbeams.svg';
 import "."
-import "./css/header.css"
-import "./css/footer.css"
+
+
+
+// import "./css/affirmations.css"
 import React from "react";
 import Homepage from "./pages/Homepage";
 import Affirmation from "./pages/Affirmation";
@@ -10,13 +12,15 @@ import Motivation from "./pages/Motivation";
 import Philosophy from "./pages/Philosophy";
 import WellnessPlanner from "./pages/WellnessPlanner";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
-import Horoscope from "./components/Horoscope";
-import Login from "./components/Login";
-import SignUp from "./components/SignUp";
-import Navbar from "./components/Navbar";
-
+import Header from "./components/Header/Header";
+import Footer from "./components/Footer/Footer";
+import Horoscope from "./components/Horoscope/Horoscope";
+import Login from "./components/Login/Login";
+import SignUp from "./components/SignUp/SignUp";
+import Navbar from "./components/Navbar/Navbar";
+import Affirmations from "./components/Affirmations";
+import Mvp from "./pages/Mvp"
+import SinglePost from "./pages/SinglePost";
 
 function App() {
   return (
@@ -24,9 +28,14 @@ function App() {
       <Router>
         <Header />
         <Navbar />
+        <Affirmation />
         <Horoscope />
+        <div style={{display: "flex"}}>
         <Login />
         <SignUp />
+        </div>
+        <Mvp />
+        < SinglePost />
         <Homepage></Homepage>
         <Switch>
           <Route exact path="/home" component={Homepage} />
