+++ riseup/src/components/Affirmations.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+
+const Affirmations = () => {
+  return (
+    <div>
+    
+    </div>
+  )
+}
+
+export default Affirmations
