+++ riseup/src/components/Components.css
@@ -1,16 +1,16 @@
-.test{
+.affirm{
     background-color: aqua;
     color: white;
     border: 2px;
-}
-.header{
+} 
+ /* .header{
     background-color:pink;
     color: black;
     border: 2px;
-}
+} 
 
-.footer{
+ .footer{
     background-color:green;
     color: black;
     border: 2px;
-}
+} */
