+++ riseup/src/components/Content.js
@@ -1,10 +1,11 @@
 import React from 'react';
 // ToDo: import any other components of styles that will be needed for this component
 import './Components.css'
+
 const Content = (props) => {
 return (
     <div>
-        <h1 className="test">
+        <h1 className="affirm">
            {props.children}
         </h1>
     </div>
