+++ riseup/src/components/Footer/footer.css
@@ -1,0 +1,17 @@
+.main-footer {
+    width: 100%;
+    height: 50%;
+    padding: 0.25em;
+    background-color: rgb(47, 47, 56);
+    color: aliceblue;
+    display: grid;
+    place-content: center;
+  } 
+
+  .list-unstyled {
+      padding-left: 10%;
+  }
+
+  .footTag {
+      color: lightblue;
+  }
