--- riseup/src/components/Header.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const Header = () => {
-  return (
-    <div>
-      <header>
-      <h1>Rise Up</h1>
-    </header>
-    </div>
-  )
-}
-
-export default Header
