+++ riseup/src/components/Header/Header.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import "../Header/header.css"
+
+const Header = () => {
+  return (
+    <div>
+      <header>
+      <h1>Rise Up</h1>
+    </header>
+    </div>
+  )
+}
+
+export default Header
