--- riseup/src/components/Navbar.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-import "../css/navbar.css"
-
-
-const NavBar = () => {
-  return (
-    <div>
-      
-    </div>
-  );
-};
-
-export default NavBar;
