+++ riseup/src/components/Navbar/navbar.css
@@ -1,0 +1,41 @@
+* {
+  box-sizing: border-box;
+  margin: 0;
+}
+
+.nav {
+  width: 100%;
+  height: 70px;
+  background-color: yellowgreen;
+  display: flex;
+  justify-content: flex-end;
+}
+
+.link {
+  display: flex;
+  padding: 10px;
+  text-decoration: none;
+  font-size: 20px;
+  /* color: black; */
+}
+
+.topLeft {
+  display: flex;
+  padding: 0.5% 65% 1%;
+}
+
+.logoImg {
+  border-radius: 30%;
+  width: 60px;
+  height: 60px;
+  object-fit: cover;
+  display: flex;
+}
+
+/* .navSearchIcon {
+  font-size: 18px;
+  color: #666;
+  cursor: pointer;
+  margin-left: 15px;
+  display: flex;
+} */
