+++ riseup/src/components/Posts/Post.js
@@ -1,0 +1,52 @@
+import React from "react";
+import "../Posts/post.css";
+
+
+const Post = () => {
+  return (
+      <div className="posts">
+    <div className="post">
+      <div className="postInfo">
+        <div className="postCats">
+        <span className="postTitle">Subject 1</span>
+          <hr/>
+          <span className="postDate">1 hour ago</span>
+          <p className="postSub">Lorem ipsum dolor sit amet consectetur adipisicing elit. Quaerat, maiores iusto perspiciatis, minus vel veniam debitis magni praesentium, excepturi nisi sint. Ab aut numquam temporibus, facilis vitae illum quae quibusdam.lit. Quaerat, maiores iusto perspiciatis, minus vel veniam debitis magni praesentium, excepturi nisi sint. Ab aut numquam temporibus, facilis vitae illum quae quibusdam.</p>
+        </div>
+      </div>
+    </div>
+    <div className="post">
+      <div className="postInfo">
+        <div className="postCats">
+        <span className="postTitle">Subject 2</span>
+          <hr/>
+          <span className="postDate">1 hour ago</span>
+          <p className="postSub">Lorem ipsum dolor sit amet consectetur adipisicing elit. Quaerat, maiores iusto perspiciatis, minus vel veniam debitis magni praesentium, excepturi nisi sint. Ab aut numquam temporibus, facilis vitae illum quae quibusdam.lit. Quaerat, maiores iusto perspiciatis, minus vel veniam debitis magni praesentium, excepturi nisi sint. Ab aut numquam temporibus, facilis vitae illum quae quibusdam.</p>
+        </div>
+      </div>
+    </div>
+    <div className="post">
+      <div className="postInfo">
+        <div className="postCats">
+          <span className="postTitle">Subject 3</span>
+          <hr/>
+          <span className="postDate">1 hour ago</span>
+          <p className="postSub">Lorem ipsum dolor sit amet consectetur adipisicing elit. Quaerat, maiores iusto perspiciatis, minus vel veniam debitis magni praesentium, excepturi nisi sint. Ab aut numquam temporibus, facilis vitae illum quae quibusdam.lit. Quaerat, maiores iusto perspiciatis, minus vel veniam debitis magni praesentium, excepturi nisi sint. Ab aut numquam temporibus, facilis vitae illum quae quibusdam.</p>
+        </div>
+      </div>
+    </div>
+    <div className="post">
+      <div className="postInfo">
+        <div className="postCats">
+        <span className="postTitle">Subject 4</span>
+          <hr/>
+          <span className="postDate">1 hour ago</span>
+          <p className="postSub">Lorem ipsum dolor sit amet consectetur adipisicing elit. Quaerat, maiores iusto perspiciatis, minus vel veniam debitis magni praesentium, excepturi nisi sint. Ab aut numquam temporibus, facilis vitae illum quae quibusdam.lit. Quaerat, maiores iusto perspiciatis, minus vel veniam debitis magni praesentium, excepturi nisi sint. Ab aut numquam temporibus, facilis vitae illum quae quibusdam.</p>
+        </div>
+      </div>
+    </div>
+    </div>
+  );
+};
+
+export default Post;
