+++ riseup/src/components/Sidebar/Sidebar.js
@@ -1,0 +1,18 @@
+import React from "react";
+import "../Sidebar/sidebar.css";
+
+const Sidebar = () => {
+  return (
+      <div className="sidebar">
+      <div className="sidebarItem">
+        <span >
+            <h1 className="sidebarTitle">Important Links</h1></span>
+        <a className="sideBtn">Crisis Link</a>
+        <a className="sideBtn">Support Group</a>
+        <a className="sideBtn">Therapist</a>
+      </div>
+      </div>
+  );
+};
+
+export default Sidebar;
