--- riseup/src/css/header.css
@@ -1,10 +1,0 @@
-header{
-  font-size: 100px;
-  background-color: plum;
-  width: calc(100% - 2px);
-  padding: 0 0.25em;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  place-content: center;
-}
