+++ riseup/src/css/mvp.css
@@ -1,0 +1,4 @@
+.mvp{
+  display: flex;  
+  background-color: rgb(201, 179, 209);
+}
