+++ riseup/src/pages/Affirmation.js
@@ -1,14 +1,11 @@
-import React from 'react';
-import Content from '../components/Content';
+import React from "react";
+import Content from "../components/Content";
 
-
-const Affirmation = () =>{
-    return (
-        <div>
-          
-            <Content>Affirmation</Content>
-          
-        </div>
-    )
-} 
+const Affirmation = () => {
+  return (
+    <div>
+      <Content>Affirmation</Content>
+    </div>
+  );
+};
 export default Affirmation;
